Added output of distance, and batch file
Distance is from search origin to station. To accomplish this I added lat lon inputs into dump_output, created a new function to calculate distance called longstanding, and imported math library. 

The bat file is a way to quickly run the code on windows. Similar to a .sh file on Linux.